AP_InertialSensor.cpp: Disable gyro calibration at start-up
This is for taking off on a rocking platform.
Accelerometer and compass are not calibrated at start-up.
Gyroscope is calibrated automatically at start-up by default.
If the uav is rocking when gyro is in the calibration procedure,
the procedure will fail and _gyro_offset will be set to zeros.
GCS will display 'Bad Gyro Health' all the time.
I'm not sure if the uav can take off in this condition.

It's safe to disable gyro calibration at start-up. The system
will load last _gyro_offset from eeprom, thus gyro will not
be influenced by the vibration.

TODO: not sure if anything relevant to pos/vel(e.g. EKF origin)
will be affected by the horizontal movement of the platform.