Call gcs_send_text() when find target in AUTO
This does not print anything on the hud on GCS as I expected.
It should have printed message in 'Messages' tab under the
hud, but it's not very convenient and I did not notice that.
But a message(MSG) is recorded in log files each time this
function is called, so it's still useful.