control_visualnav: Reorganize code due to logic flaw
The position of decision making and lifebuoy delivery is incorrect.
'target_in_image' and 'coord_near_center' conditions are not ensured.
Normally this has no problem because these 2 conditions should have
been satisfied now that copter has descended to an alt of 2m; but
we can't guarantee for the 'landing on moving platform' situation,
especially considering the real-time of program run on NanoPi is
terrible...