Record the time that target coords stay unchanged
From log files of recent flight test, I found that sometimes
the coords stay unchanged for a few seconds. I guess the program
running on NanoPi gets stuck sometimes.

TODO: I have to clear the outdated coords, or the quad will
fly away. If the quad is really loitering very steadily, it
will not cause problem to clear the coords.