do_parachute(): send signal to NanoPi and set 'nanopi_target'
We call do_parachute() when we want to start using visual info.
So we send signal to NanoPi to tell it which program to run:
nanopi_target = 1, look for delivery target;
nanopi_target = 2, look for landing target;
nanopi_target = 0, NanoPi will do nothing.

Also we set 'nanopi_target' and will send it to NanoPi through uartD
regularly to guarantee that NanoPi always executes the correct
program.